[node-notifier] Fix file structure, drop declare modules
