+++ types/node-notifier/v5/notifiers/notifysend.d.ts
@@ -1,0 +1,24 @@
+import notifier = require("node-notifier");
+
+declare class NotifySend {
+    constructor(option?: notifier.Option);
+    notify(notification?: NotifySend.Notification, callback?: notifier.NotificationCallback): NotifySend;
+}
+
+declare namespace NotifySend {
+    interface Notification {
+        title?: string | undefined;
+        message?: string | undefined;
+        icon?: string | undefined;
+        /** Specifies the urgency level  (low,  normal,  critical). */
+        urgency?: string | undefined;
+        /** Specifies  the  timeout  in  milliseconds at which to expire the notification */
+        time?: number | undefined;
+        /** Specifies the notification category */
+        category?: string | undefined;
+        /** Specifies basic extra data to pass. Valid types are int, double, string and byte. */
+        hint?: string | undefined;
+    }
+}
+
+export = NotifySend;
